중복되는 dato.ts 정리
